lastik data from 2022.01.03_looking at subset data
Today I looked at the data for the objects from the subset data. So far
everything looks good, I need to be careful not to get too sucked into
analysis here however as I still need to dissect this data by doing a
binary subtraction from the image and looked at everything in there with
multiple channels. This data simply is EDA to help me understand if this
step is working.